packages/backend/src/services/yolo-service/requirements.txt
ultralytics
opencv-python
flask
flask-cors
pillow
numpy
torch
torchvision
scikit-image
